refactor(confirm-order): adjust code with components
